refactor: abstracted utility functions buildProejctCache and fetchCacheData from project data service
